documented and modularised
added comments and created a different module for functions
